# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestor Recursos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api/v1"
        }
    ],
    "paths": {
...
        "/recourses/{recourse_id}/progress?page=1": {
            "get": {
                "tags": [
                    "Progress"
                ],
                "summary": "List History progress of recourse",
                "description": "List History progress of recourse",
                "operationId": "getProgress",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of pagination",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "identificador": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "avanzadoHasta": {
                                                        "type": "string",
                                                        "example": "0"
                                                    },
                                                    "realizado": {
                                                        "type": "string",
                                                        "example": "16"
                                                    },
                                                    "pendiente": {
                                                        "type": "string",
                                                        "example": "126"
                                                    },
                                                    "fecha": {
                                                        "type": "string",
                                                        "example": "2024-01-03"
                                                    },
                                                    "comentario": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "esUltimoRegistro": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "total": {
                                                        "type": "string",
                                                        "example": "142"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag"
                                                },
                                                "currentPage": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "perPage": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "totalPages": {
                                                    "type": "number",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=4"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recourses/{recourse_id}/progress": {
            "post": {
                "tags": [
                    "Progress"
                ],
                "summary": "Save a Progress for a Recourse",
                "description": "Save a Progress for a Recourse",
                "operationId": "postProgress",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "advanced",
                                    "date",
                                    "comment"
                                ],
                                "properties": {
                                    "advanced": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "advanced": "01:23:12",
                                    "date": "2024-12-31",
                                    "comment": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "avanzadoHasta": {
                                                    "type": "string",
                                                    "example": "0"
                                                },
                                                "realizado": {
                                                    "type": "string",
                                                    "example": "16"
                                                },
                                                "pendiente": {
                                                    "type": "string",
                                                    "example": "126"
                                                },
                                                "fecha": {
                                                    "type": "string",
                                                    "example": "2024-01-03"
                                                },
                                                "comentario": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "esUltimoRegistro": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "total": {
                                                    "type": "string",
                                                    "example": "142"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recourses/{recourse_id}/progress/{progress_id}": {
            "delete": {
                "tags": [
                    "Progress"
                ],
                "summary": "Delete a Progress for a Recourse",
                "description": "Delete a Progress for a Recourse",
                "operationId": "deleteProgress",
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 202
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "avanzadoHasta": {
                                                    "type": "string",
                                                    "example": "0"
                                                },
                                                "realizado": {
                                                    "type": "string",
                                                    "example": "16"
                                                },
                                                "pendiente": {
                                                    "type": "string",
                                                    "example": "126"
                                                },
                                                "fecha": {
                                                    "type": "string",
                                                    "example": "2024-01-03"
                                                },
                                                "comentario": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "esUltimoRegistro": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "total": {
                                                    "type": "string",
                                                    "example": "142"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}